+++ src/global_variables.c
@@ -11,4 +11,4 @@ pthread_mutex_t mutex;
 
 /* The number of cores plus one is required because 
 you have to take into account the sum of all the */
-long numCoresPlusOne;
+size_t numCoresPlusOne;
